Reorder hyperref package
Ordering of LaTeX packages is notoriously difficult, as a naive beginner you are inclined to just C-style include all you need. This poses the problem that packages loaded later overwrite commands used by other packages. Hyperref is a notorious package, "it has the nasty property that sometimes it conflicts when it is put earlier, sometimes when it is put later than other widely used packages" (Charles Stewart on TeX Stackexchange). Reading the documentation of hyperref, this should be the default order as "Make sure it comes last of your loaded packages, to give it a fighting chance of not being over-written, since its job is to redefine many LATEX commands".